Performance fix for count_dates.
I got this subtraction backward. Instead of counting updates starting
within the last week, it counted updates going back twice as long as to
the start date. But since the computed number of weeks was negative, the
extra-counted updates cancelled out and the result was still correct.
Hooray?

I would never have figured this out without the Werkzeug profiler:

http://werkzeug.pocoo.org/docs/0.14/contrib/profiler/

which told me generate_dates was entered thousands of times. That
function should only yield approximately as many times as the page size
(50), plus a few. With this bug fixed, testing the same configuration
shows 58 calls to generate_dates.