Fix timezone math for recent/next update dates.
If today is during DST and the start date was not during DST, then
for one hour after an update was supposed to occur, the max-age and
Expires headers would be in the past and the feed wouldn't add the new
item yet. However the ETag would have already changed, leading to bad
cache contents.

Just match the localize+combine pattern from generate_dates in
recent_dates and everything works out better.