Get Poetry from poetry2nix overlay, not nixpkgs
At least as of nixpkgs version 20.09.2812.f18fbc33930, using the nixpkgs
version of Poetry together with the poetry2nix repo's version of
poetry2nix results in an attempt to build pytest, which fails for some
reason.

I'm guessing this won't be an issue any more once 21.03 is out, and I'll
be able to just use the nixpkgs version of both tools.